# Optimize speed of deletion of old content metadata prior to new import