<!--
 1. Following guidance below, replace …'s with your own words
 2. After saving the PR, tick of completed checklist items
 3. Skip checklist items that are not applicable or not necessary
 4. Delete instruction/comment blocks
-->

### Summary
<!--
 * description of the change
 * manual verification steps performed
 * screenshots if the PR affects the UI
-->

Deleting old content metadata before importing a new version of a channel was super slow. This PR improves the performance in 2 ways:

1. Uses manual SQL queries to delete from specific tables, rather than doing Pythonland cascade deletion.
2. Don't delete data that's going to be overwritten efficiently with REPLACE anyway.

**Times using KA (en) channel:**

- Before: 143.8s to delete, 15.9s to import, total 159.7s
- After: 5.5s to delete, 18.5s to import, total 24s

Speedup: 26x on deletion, 6.7x overall for import updates

### Reviewer guidance
<!--
 * how can a reviewer test these changes?
 * are there any risky areas that deserve extra testing
-->

Using a primary database (db.sqlite3) that has an old version of a content database imported into it (e.g. [here's one for KA v10](https://slack-files.com/T0KT5DC58-FBGV1RECC-3b258244f3)), try importing the new version of channel. Watch the logs, and take into account the fact that a large portion of the time may be spent on downloading the new database file.

- The time from when "Older version X of channel..." is shown until "Importing ContentTag data" is shown is the deletion time. This is the part that should be much faster.
- The time from when "Importing ContentTag data" is shown to when "Setting availability of ContentNode objects with children for level 1" is shown is the remainder of the metadata import and annotation (this part shouldn't change much between old versions and this version).

----

### Contributor Checklist

- [x] Contributor has fully tested the PR manually
- [x] PR has the correct target branch and milestone
- [x] PR has 'needs review' or 'work-in-progress' label
- [x] If PR is ready for review, a reviewer has been added. (Don't use 'Assignees')
- [ ] If there are any front-end changes, before/after screenshots are included
- [ ] If this is an important user-facing change, PR or related issue has a 'changelog' label

### Reviewer Checklist

- [ ] Automated test coverage is satisfactory
- [ ] Reviewer has fully tested the PR manually
- [ ] PR has been tested for [accessibility regressions](http://kolibri-dev.readthedocs.io/en/develop/manual_testing.html#accessibility-a11y-testing)
- [ ] External dependencies files were updated (`yarn` and `pip`)
- [ ] Documentation is updated
- [ ] Link to diff of internal dependency change is included
- [ ] CHANGELOG.rst is updated for high-level changes
- [ ] Contributor is in AUTHORS.rst