@radinamatic 2018-07-03
Waiting for the asset to start testing! 👍