@radinamatic 2018-07-03
> Altro que volare... ❤️ 

PEX on Ubuntu 16.04, VM with 2GB RAM, single core:

- 18s to delete
- 41s to import

```INFO 2018-07-03 03:08:00,600 ping Attempting a ping.
INFO 2018-07-03 03:08:00,648 ping data: {"instance_id": "c394cef1b732b2e5761db4bc69258993", "platform": "Linux-4.4.0-128-generic-x86_64-with-Ubuntu-16.04-xenial", "database_id": "a50ab60977814c40b16870f5da2b8a4c", "version": "0.10.0.post1", "system_id": "", "mode": "LE-testingHD-U16043", "language": "en", "sysversion": "2.7.12 (default, Dec  4 2017, 14:50:18) \n[GCC 5.4.0 20160609]", "uptime": 0, "node_id": "baa72ad1f5714b23762a"}
INFO 2018-07-03 03:08:00,781 utils Started new MESSAGEPROCESSOR thread ID#lake-pasta-ack-july
INFO 2018-07-03 03:08:00,781 utils Started new SCHEDULER thread ID#washington-lima-utah-nebraska
INFO 2018-07-03 03:08:00,785 utils Started new WORKERMESSAGEHANDLER thread ID#jig-lithium-spaghetti-fillet
INFO 2018-07-03 03:08:01,232 ping Ping succeeded! (response: {}) Sleeping for 1440.0 minutes.
INFO 2018-07-03 03:09:27,796 importchannel Downloading data for channel id 1ceff53605e55bef987d88e0908658c5
INFO 2018-07-03 03:09:35,471 channel_import Older version 10 of channel 1ceff53605e55bef987d88e0908658c5 already exists in database; removing old entries so we can upgrade to version 11
INFO 2018-07-03 03:09:57,587 channel_import Importing ContentTag data
INFO 2018-07-03 03:09:57,592 channel_import Importing ContentNode_has_prerequisite data
INFO 2018-07-03 03:09:57,594 channel_import Importing ContentNode_related data
INFO 2018-07-03 03:09:57,594 channel_import Importing ContentNode_tags data
INFO 2018-07-03 03:09:57,595 channel_import Importing ContentNode data
INFO 2018-07-03 03:10:04,243 channel_import Importing Language data
INFO 2018-07-03 03:10:04,245 channel_import Importing File data
INFO 2018-07-03 03:10:26,887 channel_import Importing LocalFile data
INFO 2018-07-03 03:10:28,522 channel_import Importing AssessmentMetaData data
INFO 2018-07-03 03:10:28,952 channel_import Importing ChannelMetadata data
INFO 2018-07-03 03:10:31,519 annotation Setting availability of File objects based on LocalFile availability
INFO 2018-07-03 03:10:36,025 annotation Setting availability of non-topic ContentNode objects based on File availability
INFO 2018-07-03 03:10:38,093 annotation Setting availability of ContentNode objects with children for 5 levels
INFO 2018-07-03 03:10:38,094 annotation Setting availability of ContentNode objects with children for level 5
INFO 2018-07-03 03:10:38,249 annotation Setting availability of ContentNode objects with children for level 4
INFO 2018-07-03 03:10:38,296 annotation Setting availability of ContentNode objects with children for level 3
INFO 2018-07-03 03:10:38,300 annotation Setting availability of ContentNode objects with children for level 2
INFO 2018-07-03 03:10:38,307 annotation Setting availability of ContentNode objects with children for level 1```