@radinamatic 2018-07-03
For comparison, the whole process in 0.10b6 PEX took 15 minutes:

```INFO 2018-07-03 03:26:55,360 importchannel Downloading data for channel id 1ceff53605e55bef987d88e0908658c5
INFO 2018-07-03 03:27:04,300 channel_import Older version 10 of channel 1ceff53605e55bef987d88e0908658c5 already exists in database; removing old entries so we can upgrade to version 11
ERROR 2018-07-03 03:38:10,908 exception The request's session was deleted before the request completed. The user may have logged out in a concurrent request, for example.
INFO 2018-07-03 03:41:09,141 channel_import Importing ContentTag data
INFO 2018-07-03 03:41:20,440 channel_import Importing ContentNode_has_prerequisite data
INFO 2018-07-03 03:41:20,761 channel_import Importing ContentNode_related data
INFO 2018-07-03 03:41:20,785 channel_import Importing ContentNode_tags data
INFO 2018-07-03 03:41:20,807 channel_import Importing ContentNode data
INFO 2018-07-03 03:41:31,090 channel_import Importing Language data
INFO 2018-07-03 03:41:31,116 channel_import Importing File data
INFO 2018-07-03 03:42:14,689 channel_import Importing LocalFile data
INFO 2018-07-03 03:42:18,336 channel_import Importing AssessmentMetaData data
INFO 2018-07-03 03:42:21,676 channel_import Importing ChannelMetadata data
INFO 2018-07-03 03:42:32,727 annotation Setting availability of File objects based on LocalFile availability
INFO 2018-07-03 03:42:39,198 annotation Setting availability of non-topic ContentNode objects based on File availability
INFO 2018-07-03 03:42:43,658 annotation Setting availability of ContentNode objects with children for 5 levels
INFO 2018-07-03 03:42:43,668 annotation Setting availability of ContentNode objects with children for level 5
INFO 2018-07-03 03:42:43,822 annotation Setting availability of ContentNode objects with children for level 4
INFO 2018-07-03 03:42:43,868 annotation Setting availability of ContentNode objects with children for level 3
INFO 2018-07-03 03:42:43,881 annotation Setting availability of ContentNode objects with children for level 2
INFO 2018-07-03 03:42:43,887 annotation Setting availability of ContentNode objects with children for level 1```