@jamalex 2018-07-03
Wooooooooo! 14m down to 22s for the deletion portion!