@benjaoming 2018-07-03
> Uses manual SQL queries to delete from specific tables, rather than doing Pythonland cascade deletion.

This doesn't happen if you call `delete()` on a QuerySet. It happens if you call `delete()` on an object instance.

I am just reciting this out of memory, but also a distinct feeling that I've done it before :) If you don't mind, could you try and see if things are still working if you make this change?

```python
delete_related_template = """
    DELETE FROM {table}
        WHERE {fk_field} IN (
            SELECT id FROM {cn_table} WHERE tree_id = '{tree_id}'
        )
"""
```

Becomes:

```python
function delete_ids_by_fk_field_in_tree(DeleteFromModel, ChannelModel, fk_field, tree_id):
    # Construct the IN query
    kwargs={fk_field + "__in": ChannelModel.objects.filter(tree_id=tree_id).values('id')}

    # Construct and execute the DELETE query
    DeleteFromModel.filter(**kwargs).delete()
```