@rtibbles 2018-07-03
@benjaoming note that we can't do this in Django, this was moved from Django to SQLAlchemy in #3538 in order to allow the deletion to happen within the same transaction as the channel import process that happens within SQLAlchemy (as I have yet to find a way to make Django and SQLAlchemy share a transaction).

The slowness that @jamalex was referring to was mostly in the SQLAlchemy reflection of the relationships defined within Django (which when triggering its cascade behaviour would somehow do something more inefficient than the Django defined cascade behaviour).

The only open question for this is whether translating these statements into SQLAlchemy core syntax as opposed to raw SQL - but ultimately the core syntax is just a Pythonic way of constructing SQL expressions.