@benjaoming 2018-07-05
> ultimately the core syntax is just a Pythonic way of constructing SQL expressions

Yes, agreed -- I don't know about advantages of SQLAlchemy, I wasn't aware that these components were living there.

In Django, the advantage would be about controlling model and field references in Python, as opposed to coding them into SQL strings + Being DB agnostic + Also, custom managers may influence querysets in some global way.

Anyhow, alright, seems like we'll stick to the raw SQL :)